Updates issue 111
Generated IDocLibrary based on new meta model.

This resulted in a lot of compilation errors throughout the project, because the new MA concept has finally been introduces.

I have fixed the problems where it was easily possible, but also added a lot of NotImplementedExceptions and set a bunch of tests to Ignore.