Fixed Issue 128.
So far, for importing the standard Core Component Library, the VIENNA Add-In required to be online in order to retrieve the latest set of Libraries. Now, a standard set of Core Component Libraries is delivered with the VIENNA Add-In Installation package which is available for import if no Internet connection is available.